fix playground
