+++ playground/react/src/App.style.ts
@@ -1,0 +1,17 @@
+// import { classes } from './App.style';
+
+import { globalStyle, style } from '@kaze-style/core';
+import { globalTheme } from '@kaze-style/themes';
+
+globalStyle(globalTheme.reset());
+
+export const classes = style({
+  button: {
+    background: 'red',
+    borderRadius: '6px',
+    padding: '10px 20px',
+    ':hover': {
+      background: 'orange',
+    },
+  },
+});
